+++ src/Main.java
@@ -6,7 +6,7 @@ public class Main {
         int n = 1;
         int k = 5;
 
-        int[] testArray = BorrowedBench.keys(k, n);
+        float[] testArray = BorrowedBench.keys(k, n);
 
         for (int i = 0; i < k; i++) {
             System.out.println(testArray[i]);
